Modifying word2vec.cpp in more OOP style, while finding possible bugs. Still in progress.
Summary:
The class Word2Vec is defined as a subclass of WordEmbed, which is currently empty but given as a placeholder for the further refactoring. All the global variables in the previous implementation are now member variables of Word2Vec, except for the constant variables used to set the length of the array like objects. The argument parser routine is also modified so that the command line options are directly fed to the constructor of the class.

Test Plan: None

Reviewers: modori, wavewave

Reviewed By: wavewave

Differential Revision: https://uphere.phacility.com/D242